Update regarding more yes options
Here is an update that includes more options for the user to say "yes" to the program when the program asks whether the user needs to change their password.

Now, when prompted with that question, the user can also reply with other versions of "yes".  For example, they can instead write "yah".